removed and renamed the clustering class and script
